* feat: GET all Tenants endpoint

* refactor: update swagger

* feat: Delete buddies endpoint

* Bug: Unable to Publish Property (keeps on asking for profile to be completed)

* make property_id nullable

* fix: update dev environment

* fix: update pipeline

* fix: update rsync flag

* Update api/landlords endpoint (#21)

* Create Endpoint to return the buddies of the landlord

* Assign a random property ID in portfolio import

* Create and endpoint to delete buddies (single and multiple selection)

* Create and endpoint to delete buddies (single and multiple selection)

* Assign 8 digit random alphanumeric  property ID in portfolio import

+++ app/Services/EstateService.js
@@ -566,7 +566,14 @@ class EstateService {
    *
    */
   static async publishEstate(estate) {
-    await CompanyService.validateUserContacts(estate.user_id)
+    const User = use('App/Models/User')
+    const user = await User.query()
+      .where('id',  estate.user_id)
+      .first()
+    if (!user) return
+    if (user.company_id != null){
+      await CompanyService.validateUserContacts(estate.user_id)
+    }
     await props({
       delMatches: () => Database.table('matches').where({ estate_id: estate.id }).delete(),
       delLikes: () => Database.table('likes').where({ estate_id: estate.id }).delete(),
